Activity Manager

+++ trunk/plugins/tools/ActivityManager-portlet/docroot/WEB-INF/service/com/liferay/lmssa/model/ActManAuditClp.java
@@ -144,7 +144,7 @@ public class ActManAuditClp extends BaseModelImpl<ActManAudit>
 			setState(state);
 		}
 
-		Long number = (Long)attributes.get("number");
+		Integer number = (Integer)attributes.get("number");
 
 		if (number != null) {
 			setNumber(number);
@@ -239,11 +239,11 @@ public class ActManAuditClp extends BaseModelImpl<ActManAudit>
 		_state = state;
 	}
 
-	public long getNumber() {
+	public int getNumber() {
 		return _number;
 	}
 
-	public void setNumber(long number) {
+	public void setNumber(int number) {
 		_number = number;
 	}
 
@@ -432,6 +432,6 @@ public class ActManAuditClp extends BaseModelImpl<ActManAudit>
 	private Date _start;
 	private Date _end;
 	private String _state;
-	private long _number;
+	private int _number;
 	private BaseModel<?> _actManAuditRemoteModel;
 }
